Changed FileCache to allow for asynchronous reading.
Maintains backwards compatibility with any filetype that uses
synchronous reading, as long as it never returns -2 as error code.
-2 is used to indicate that an async read is ongoing. (Note that
the error code can be changed to anything else, if necessary, it
was chosen simply because this is the code returned by libssh)

Tested only to make sure it doesn't break synchronous reads, since
no async filetypes exist yet (working on this for sftp).